# backend/price_tracker/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервер для відстеження цін на товари.",
        "title": "Price Tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Перевіряє пошту/пароль та повертає JWT-токен",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Авторизація користувача",
                "parameters": [
                    {
                        "description": "Дані для входу",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішний вхід, повертає токен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неправильний пароль або логін",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Створює новий акаунт користувача в базі даних",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Реєстрація нового користувача",
                "parameters": [
                    {
                        "description": "Дані для реєстрації",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.RegisterInput"
                        }
                    }
                ],
                "responses": {
                    "21": {
                        "description": "Успішна реєстрація",
                        "schema": {
                            "type": "stage_success"
                        }
                    },
                    "400": {
                        "description": "Помилка у запиті або дублікат Email",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Повертає список усіх товарів, які відстежує поточний авторизований користувач",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Отримати товари користувача",
                "responses": {
                    "200": {
                        "description": "Список товарів",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/price-tracker_internal_models.Product"
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизовано",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Додає посилання на товар до акаунта поточного користувача. Якщо товар новий — створює його.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Додати товар для відстеження",
                "parameters": [
                    {
                        "description": "URL товару",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.AddProductInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Товар успішно додано",
                        "schema": {
                            "$ref": "#/definitions/price-tracker_internal_models.Product"
                        }
                    },
                    "400": {
                        "description": "Некоректний URL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизовано",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Видаляє зв'язок між поточним користувачем та товаром за його ID. Якщо товар більше ніхто не відстежує, він повністю видаляється з бази даних.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Видалити товар із відстеження",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID товару, який потрібно видалити",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Товар успішно видалено з відстеження",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некоректний ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизовано",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Товар або користувача не знайдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_handlers.AddProductInput": {
            "type": "object",
            "required": [
                "name",
                "url"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.LoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.RegisterInput": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "price-tracker_internal_models.Product": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "product_last_price": {
                    "type": "number"
                },
                "product_name": {
                    "type": "string"
                },
                "product_url": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введіть токен у форматі: Bearer \u003cваш_токен\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}